pipeline/pyproject.toml
[build-system]
requires = ["setuptools>=75.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "noir-canvas-pipeline"
version = "0.1.0"
description = "AI art generation pipeline for Noir Canvas"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "pydantic>=2.11",
    "pydantic-settings>=2.9",
    "httpx>=0.28",
    "Pillow>=11.2",
    "pyyaml>=6.0",
    "google-genai>=1.14",
    "websockets>=15.0",
    "jsonschema>=4.23",
    "numpy>=1.26",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3",
    "pytest-asyncio>=0.26",
    "ruff>=0.11",
    "mypy>=1.15",
]

[project.scripts]
pipeline = "pipeline.cli:cli"

[tool.setuptools.packages.find]
where = ["src"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
